First commit second deliverable
For this first commit I change some css in the header and home files, I feel more confident with this arrangement. Also I start to add the new products page, with the menu/filters and started the products list. I gave format to the file with the list of products with prettier